Revert decimal to double for simplicity

+++ WalletApp.Application/DTOs/AccountInfoDTO.cs
@@ -2,7 +2,7 @@ using WalletApp.Domain.Entities;
 
 namespace WalletApp.Application.DTOs
 {
-    public record AccountInfoDTO(decimal CardBalance, string PaymentDueStatus, string DailyPointsMessage, 
+    public record AccountInfoDTO(double CardBalance, string PaymentDueStatus, string DailyPointsMessage, 
         IEnumerable<TransactionPreviewDTO> Transactions)
     {
         public AccountInfoDTO(AccountInfo accountInfo, IEnumerable<TransactionPreviewDTO> transactions)
